Moving frame_audio to a util to clean up API
