+++ hearkit/baseline.py
@@ -9,8 +9,9 @@ from typing import Tuple
 
 import librosa
 import torch
-import torch.nn.functional as F
 from torch import Tensor
+
+from hearkit.util import frame_audio
 
 # Default hop_size in milliseconds
 HOP_SIZE = 25
@@ -20,6 +21,10 @@ BATCH_SIZE = 512
 
 
 class RandomProjectionMelEmbedding(torch.nn.Module):
+    """
+    Baseline audio embedding model. This model creates mel frequency spectrums with
+    256 mel-bands, and then performs a projection to an embedding size of 4096.
+    """
 
     # sample rate and embedding sizes are required model attributes for the HEAR API
     sample_rate = 44100
@@ -89,53 +94,6 @@ def load_model(model_file_path: str = "") -> torch.nn.Module:
         raise NotImplementedError("Loading model weights not implemented yet")
 
     return model
-
-
-def frame_audio(
-    audio: Tensor, frame_size: int, hop_size: float, sample_rate: int
-) -> Tuple[Tensor, Tensor]:
-    """
-    Slices input audio into frames that are centered and occur every
-    sample_rate * hop_size samples. We round to the nearest sample.
-
-    Args:
-        audio: input audio, expects a 2d Tensor of shape:
-            (batch_size, num_samples)
-        frame_size: the number of samples each resulting frame should be
-        hop_size: hop size between frames, in milliseconds
-        sample_rate: sampling rate of the input audio
-
-    Returns:
-        - A Tensor of shape (batch_size, num_frames, frame_size)
-        - A 1d Tensor of timestamps corresponding to the frame
-        centers.
-    """
-
-    # Zero pad the beginning and the end of the incoming audio with half a frame number
-    # of samples. This centers the audio in the middle of each frame with respect to
-    # the timestamps.
-    audio = F.pad(audio, (frame_size // 2, frame_size - frame_size // 2))
-    num_padded_samples = audio.shape[1]
-
-    frame_number = 0
-    frames = []
-    timestamps = []
-    frame_start = 0
-    frame_end = frame_size
-    while True:
-        frames.append(audio[:, frame_start:frame_end])
-        timestamps.append(frame_number * hop_size)
-
-        # Increment the frame_number and break the loop if the next frame end
-        # will extend past the end of the padded audio samples
-        frame_number += 1
-        frame_start = int(round(sample_rate * frame_number * hop_size / 1000))
-        frame_end = frame_start + frame_size
-
-        if not frame_end <= num_padded_samples:
-            break
-
-    return torch.stack(frames, dim=1), torch.tensor(timestamps)
 
 
 def get_timestamp_embeddings(
